% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{senate}
\alias{senate}
\title{Senate data from ..}
\format{A data frame with 147 rows (votes) and 9 variables (parties).
#'@examples
dim(senate)}
\source{
\url{}
}
\usage{
data(senate)
}
\description{
A dataset containing the
each row is, each col is, data is either, yes, no, NA
}
\keyword{datasets}
