% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoltzMM.R
\docType{data}
\name{senate}
\alias{senate}
\title{Senate voting data from the 45th Australian Parliament.}
\format{A data frame with 147 rows (votes) and 9 variables (parties).}
\source{
\url{www.aph.gov.au/Parliamentary_Business/Statistics/Senate_StatsNet/General/divisions}
}
\usage{
data(senate)
}
\description{
A dataset he data from the first sitting of the Senate of the 45th
Australian Parliament, until the final sitting of the year 2016. The first division during
this period was conducted on the 31st of August 2016, and the last division was performed
on the 1st of December 2016. In total, 147 divisions were performed during this period.

Each row represents a division(vote), each column is a party or independent.
Data is either "Yes" or "No" depending on the vote. Absences and abstentions are left as NA.
See \url{https://hal.archives-ouvertes.fr/hal-01927188v1} for details of data preparation.
}
\examples{
dim(senate)
}
\author{
Jessica J. Bagnall
}
\keyword{datasets}
